RING-28615 Split client's GET/PUT/DELETE in several files
