+++ lib/hdclient.js
@@ -1,13 +1,13 @@
 'use strict'; // eslint-disable-line strict
 
 const http = require('http');
-const assert = require('assert');
 const werelogs = require('werelogs');
 
 const config = require('./config');
-const protocol = require('./protocol');
 const keyscheme = require('./keyscheme');
-const httpUtils = require('./http_utils');
+const doGET = require('./http_get');
+const doPUT = require('./http_put');
+const doDELETE = require('./http_delete');
 
 
 class HyperdriveClient {
@@ -78,24 +78,6 @@ class HyperdriveClient {
             this.logging.newRequestLogger();
     }
 
-    /**
-     * Construct option object to use for next store query
-     *
-     * @param {String} hostname to contact
-     * @param {Number} port to contact on
-     * @param {String} key to target
-     * @returns {Object} HTTP client request object filled with common options
-     */
-    _getCommonStoreRequestOptions(hostname, port, key) {
-        return {
-            hostname,
-            port,
-            path: `${protocol.specs.STORAGE_BASE_URL}/${key}`,
-            agent: this.httpAgent,
-            protocol: 'http:',
-        };
-    }
-
     /** **************************************************************
      * Data backend interface
      *
@@ -118,6 +100,8 @@ class HyperdriveClient {
      * @returns {undefined}
      */
     put(stream, size, keyContext, reqUids, callback) {
+        const log = this.createLogger(reqUids);
+
         // Select hyperdrives
         const fragments = keyscheme.keygen(
             this.options.policy,
@@ -130,114 +114,38 @@ class HyperdriveClient {
 
         const rawGenKey = keyscheme.serialize(fragments);
 
-        // Split, replication or erasure coding is currently not supported
-        assert.strictEqual(fragments.nChunks, 1);
-        assert.strictEqual(fragments.nCodingParts, 0);
-
-        const log = this.createLogger(reqUids);
-        const opContext = httpUtils.makeOperationContext(fragments);
-
-        /* TODO: split & erasure coding */
-
-        /* Dispatch replication */
-        fragments.chunks.slice(0, 1).forEach((chunk, chunkId) => {
-            chunk.data.forEach((fragment, fragmentId) => {
-                const reqContext = { opContext, chunkId, fragmentId };
-                const { hostname, port, key } = fragment;
-                const contentType = protocol.helpers.makePutContentType(
-                    { data: size } /* Only 'data' payload is supported */
-                );
-                const requestOptions = this._getCommonStoreRequestOptions(
-                    hostname, port, key);
-
-                requestOptions.method = 'PUT';
-                requestOptions.headers = {
-                    ['Content-Length']: size,
-                    ['Content-Type']: contentType,
-                };
-
-                const request = httpUtils.newRequest(
-                    requestOptions, log, reqContext,
-                    this.options.requestTimeoutMs,
-                    /* callback */
-                    opCtx => {
-                        log.end();
-                        const endStatus =
-                                  opCtx.status[chunkId].statuses[fragmentId];
-                        if (endStatus.timeout) {
-                            return callback(null, rawGenKey);
-                        }
-
-                        return callback(endStatus.error, rawGenKey);
-                    });
-
-                // TODO abstract a bit stream to have a safer interface
-                stream.pipe(request);
-                stream.on('error', err => {
-                    // forward error downstream
-                    request.emit('error', err);
-                });
-            });
-        });
+        doPUT({ log, fragments, rawGenKey,
+                callback, size, stream,
+                httpAgent: this.httpAgent,
+                requestTimeoutMs: this.options.requestTimeoutMs });
     }
 
     /**
      * This sends a GET request to hyperdrives.
      * @param {String} rawKey - The key associated to the value
-     * @param {Number [] | Undefined} range - range (if any) with
-     *                                         first element the start
-     * and the second element the end
+     * @param {Number [] | Undefined} range - Range (if any) with
+     *                                        first element the start
+     *                                        and the second element the end
      * @param {String} reqUids - The serialized request id
      * @param {HyperdriveClient~getCallback} callback - callback
      * @returns {undefined}
      */
     get(rawKey, range, reqUids, callback) {
+        const log = this.createLogger(reqUids);
         let fragments;
         try {
             fragments = keyscheme.deserialize(rawKey);
         } catch (error) {
+            log.error(`Failed to deserialize key: ${rawKey}`,
+                      error.message);
             callback(error);
+            log.end();
             return;
         }
 
-        // Split, replication or erasure coding is currently not supported
-        assert.strictEqual(fragments.nChunks, 1);
-        assert.strictEqual(fragments.nCodingParts, 0);
-
-        const log = this.createLogger(reqUids);
-        const opContext = httpUtils.makeOperationContext(fragments);
-
-        /* Request first chunk all at once
-         * TODO: split? Request all? First 2 then the next when
-         *       first is consumed? Introduced a small delay (1ms)?
-         * TODO: handle range on split
-         */
-        fragments.chunks.slice(0, 1).forEach((chunk, chunkId) => {
-            [...chunk.data, ...chunk.coding].forEach((fragment, fragmentId) => {
-                const reqContext = { opContext, chunkId, fragmentId };
-                const { hostname, port, key } = fragment;
-                const requestOptions = this._getCommonStoreRequestOptions(
-                    hostname, port, key);
-
-                requestOptions.method = 'GET';
-                requestOptions.headers = {
-                    ['Accept']: protocol.helpers.makeAccept(['data', range]),
-                };
-
-                const request = httpUtils.newRequest(
-                    requestOptions, log, reqContext,
-                    this.options.requestTimeoutMs,
-                    /* callback */
-                    opCtx => {
-                        log.end();
-                        const status =
-                                  opCtx.status[chunkId].statuses[fragmentId];
-                        return callback(status.error, status.response);
-                    });
-
-                request.end();
-            });
-        });
+        doGET({ fragments, range, callback, log,
+                httpAgent: this.httpAgent,
+                requestTimeoutMs: this.options.requestTimeoutMs });
     }
 
     /**
@@ -249,49 +157,21 @@ class HyperdriveClient {
      * @returns {undefined}
      */
     delete(rawKey, reqUids, callback) {
+        const log = this.createLogger(reqUids);
         let fragments;
         try {
             fragments = keyscheme.deserialize(rawKey);
         } catch (error) {
+            log.error(`Failed to deserialize key: ${rawKey}`,
+                      error.message);
             callback(error);
+            log.end();
             return;
         }
 
-        // Split, replication or erasure coding is currently not supported
-        assert.strictEqual(fragments.nChunks, 1);
-        assert.strictEqual(fragments.nCodingParts, 0);
-
-        const log = this.createLogger(reqUids);
-        const opContext = httpUtils.makeOperationContext(fragments);
-
-        /* Send all at once */
-        fragments.chunks.forEach((chunk, chunkId) => {
-            [...chunk.data, ...chunk.coding].forEach((fragment, fragmentId) => {
-                const reqContext = { opContext, chunkId, fragmentId };
-                const { hostname, port, key } = fragment;
-                const requestOptions = this._getCommonStoreRequestOptions(
-                    hostname, port, key);
-
-                requestOptions.method = 'DELETE';
-                requestOptions.headers = {
-                    ['Content-Length']: 0,
-                    ['Accept']: protocol.helpers.makeAccept(),
-                };
-
-                const request = httpUtils.newRequest(
-                    requestOptions, log, reqContext,
-                    this.options.requestTimeoutMs,
-                    /* callback */
-                    opCtx => {
-                        log.end();
-                        const status =
-                                  opCtx.status[chunkId].statuses[fragmentId];
-                        return callback(status.error);
-                    });
-
-                request.end();
-            });
-        });
+        doDELETE({ fragments, callback, log,
+                   httpAgent: this.httpAgent,
+                   requestTimeoutMs: this.options.requestTimeoutMs });
     }
 
     /**
