+++ lib/http_utils.js
@@ -3,6 +3,8 @@
 const assert = require('assert');
 const http = require('http');
 
+const protocol = require('./protocol');
+
 /**
  * Create a new Operation Context
  *
@@ -188,8 +190,28 @@ function newRequest(options, logger, reqContext, timeoutMs, callback) {
     return request;
 }
 
+/**
+ * Construct option object to use for next store query
+ *
+ * @param  {http.HttpAgent} httpAgent to use
+ * @param {String} hostname to contact
+ * @param {Number} port to contact on
+ * @param {String} key to target
+ * @returns {Object} HTTP client request object filled with common options
+ */
+function getCommonStoreRequestOptions(httpAgent, hostname, port, key) {
+    return {
+        hostname,
+        port,
+        path: `${protocol.specs.STORAGE_BASE_URL}/${key}`,
+        agent: httpAgent,
+        protocol: 'http:',
+    };
+}
+
 
 module.exports = {
+    getCommonStoreRequestOptions,
     newRequest,
     makeOperationContext,
 };
